Add messaging to highlight when data is being uploaded to public beta servers

+++ siliconcompiler/apps/sc_configure.py
@@ -3,6 +3,24 @@ import os
 import sys
 from pathlib import Path
 
+tos_str = '''Please review the SiliconCompiler cloud beta's terms of service:
+
+https://www.siliconcompiler.com/terms-of-service
+
+In particular, please ensure that you have the right to distribute any IP which is contained in designs that you upload to the service. This public service, provided by SiliconCompiler, is not intended to process proprietary IP.
+'''
+
+def confirm_dialog(message):
+    confirmed = False
+    while not confirmed:
+        oin = input(message)
+        if (not oin) or (oin == 'n') or (oin == 'N'):
+            print('Exiting.')
+            return 1
+        elif (oin == 'y') or (oin == 'Y'):
+            confirmed = True
+    return 0
+
 def main():
     progname = "sc-configure"
     switchlist = []
@@ -30,18 +48,14 @@ def main():
 
     # If an existing config file exists, prompt the user to overwrite it.
     if os.path.isfile(cfg_file):
-        overwrite = False
-        while not overwrite:
-            oin = input('Overwrite existing remote configuration? (y/N)')
-            if (not oin) or (oin == 'n') or (oin == 'N'):
-                print('Exiting.')
-                return
-            elif (oin == 'y') or (oin == 'Y'):
-                overwrite = True
+        if confirm_dialog('Overwrite existing remote configuration?'):
+            return
 
     # If a command-line argument is passed in, use that as a public server address.
     if len(sys.argv) > 1:
         print(f'Creating remote configuration file for public server: {sys.argv[1]}')
+        if 'server.siliconcompiler.com' in sys.argv[1] and confirm_dialog(tos_str):
+            return
         with open(cfg_file, 'w') as f:
             f.write('{"address": "%s"}'%(sys.argv[1]))
         return
@@ -50,6 +64,9 @@ def main():
     srv_addr = input('Remote server address:\n').replace(" ","")
     username = input('Remote username (leave blank for public servers):\n').replace(" ","")
     user_pass = input('Remote password (leave blank for public servers):\n').replace(" ","")
+
+    if 'server.siliconcompiler.com' in srv_addr and confirm_dialog(tos_str):
+        return
 
     # Save the values to the target config file in JSON format.
     with open(cfg_file, 'w') as f:
